+++ packages/desktop/publish/changeLog.md
@@ -14,6 +14,8 @@
 
 - Fixed an issue where moving songs could result in an incorrect saved order in some cases.
 - Fixed an issue with lyric timestamp tag parsing.
+- Fixed an issue where the **Lyric Offset** feature was not functioning correctly.
+- Fixed an issue where the playback speed feature was not functioning correctly.
 - Fixed a playlist synchronization issue when playing online lists ([#225](https://github.com/any-listen/any-listen/issues/225)).
 
 ### Changed
@@ -38,6 +40,8 @@
 
 - 修复在某些情况下移动歌曲时可能导致保存的顺序不对的问题。
 - 修复歌词时间标签解析问题。
+- 修复歌词偏移功能异常的问题。
+- 修复播放速率功能异常的问题。
 - 修复播放在线列表时的播放列表同步问题（[#225](https://github.com/any-listen/any-listen/issues/225)）。
 
 ### 变更